requirement.txt
undetected-chromedriver
selenium
requests
m3u8
pycryptodome
ffmpeg-python
rich
